+++ src/shared/components/case-viewer/viewer.routing.ts
@@ -6,15 +6,16 @@ import { EventStartGuard } from '../event-guard/event-start.guard';
 import { CaseChallengedAccessRequestComponent } from './case-challenged-access-request';
 import { CaseChallengedAccessSuccessComponent } from './case-challenged-access-success';
 import { CaseSpecificAccessRequestComponent } from './case-specific-access-request';
+import { CaseReviewSpecificAccessRequestComponent } from './case-review-specific-access-request';
 import { CaseEventTriggerComponent } from './case-event-trigger/case-event-trigger.component';
 import { CasePrinterComponent } from './printer';
 import { EventTriggerResolver } from './services';
+import { CaseReviewSpecificAccessRejectComponent } from './case-review-specific-access-reject';
 import { TaskAssignedComponent } from '../event-trigger/components/task-assigned/task-assigned.component';
 import { MultipleTasksExistComponent } from '../event-trigger/components/multiple-tasks-exist/multiple-tasks-exist.component';
 import { NoTasksAvailableComponent } from '../event-trigger/components/no-tasks-available/no-tasks-available.component';
 import { TaskCancelledComponent } from '../event-trigger/components/task-cancelled/task-cancelled.component';
 import { TaskConflictComponent } from '../event-trigger/components/task-conflict/task-conflict.component';
-
 export const viewerRouting: Routes = [
   {
     path: 'print',
@@ -23,12 +24,12 @@ export const viewerRouting: Routes = [
   {
     path: 'trigger/:eid',
     resolve: {
-      eventTrigger: EventTriggerResolver
+      eventTrigger: EventTriggerResolver,
     },
     component: CaseEventTriggerComponent,
     children: editorRouting,
     canActivate: [EventStartGuard],
-    canDeactivate: [FileUploadProgressGuard]
+    canDeactivate: [FileUploadProgressGuard],
   },
   {
     path: 'task-assignment',
@@ -69,18 +70,18 @@ export const viewerRouting: Routes = [
         path: '',
         component: CaseChallengedAccessRequestComponent,
         data: {
-          title: 'Request Challenged Access'
+          title: 'Request Challenged Access',
         },
-        pathMatch: 'full'
+        pathMatch: 'full',
       },
       {
         path: 'success',
         component: CaseChallengedAccessSuccessComponent,
         data: {
-          title: 'Challenged Access Success'
-        }
-      }
-    ]
+          title: 'Challenged Access Success',
+        },
+      },
+    ],
   },
   {
     path: 'specific-access-request',
@@ -89,10 +90,30 @@ export const viewerRouting: Routes = [
         path: '',
         component: CaseSpecificAccessRequestComponent,
         data: {
-          title: 'Request Specific Access'
+          title: 'Request Specific Access',
         },
-        pathMatch: 'full'
+        pathMatch: 'full',
+      },
+    ],
+  },
+  {
+    path: 'review-specific-access-request',
+    children: [
+      {
+        path: '',
+        component: CaseReviewSpecificAccessRequestComponent,
+        data: {
+          title: 'Request Specific Access',
+        },
+        pathMatch: 'full',
+      },
+      {
+        path: 'rejected',
+        component: CaseReviewSpecificAccessRejectComponent,
+        data: {
+          title: 'Review Access Rejected'
+        }
       }
-    ]
-  }
+    ],
+  },
 ];