+++ RELEASE-NOTES.md
@@ -1,5 +1,12 @@
 ## RELEASE NOTES
 
+### Version 4.7.7-hearings-tab
+**EUI-4441** Add hearings tab in case details
+**EUI-4627** Can't access /hearings tab from URL
+
+### Version 4.7.0-EUI-4412-order-fields-CYA-page
+**EUI-4414** Field order on the CYA page changes
+
 ### Version 3.1.5-complex-type-showing-in-tab
 **EUI-3798** 'Empty' complex type showing in tab (due to Label)